Change readLine ints to longs

+++ globals.c
@@ -2,7 +2,7 @@
 
 #include "globals.h"
 
-int getIndexOfElement(char* string, int n) {
+long getIndexOfElement(char* string, long n) {
 
 	if (n < 1) {
 		return -1;
@@ -11,7 +11,7 @@ int getIndexOfElement(char* string, int n) {
 	int counter = 0; // Conta o número de dois pontos (:) encontrados (a cada momento estamos no elemento counter + 1)
 	char c;
 
-	int i;
+	long i;
 	for (i = 0; i < strlen(string); i++) {
 
 		c = string[i];
